requirements.txt
astroid==1.6.0
Django==2.0
djangorestframework==3.7.3
flake8==3.5.0
isort==4.2.15
lazy-object-proxy==1.3.1
markdown2==2.3.5
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==1.8.1
pytz==2017.3
six==1.11.0
wrapt==1.10.11
yapf==0.20.0
